HTBHF-2507 move test utils to top level of flow control directory
