+++ src/web/routes/application/flow-control/index.js
@@ -1,7 +1,9 @@
 const { getPreviousPath } = require('./get-previous-path')
+const testUtils = require('./test-utils')
 
 module.exports = {
   ...require('./state-machine'),
   ...require('./middleware'),
-  getPreviousPath
+  getPreviousPath,
+  testUtils
 }
