+++ src/web/routes/application/flow-control/middleware/handle-path-request/handle-path-request.test.js
@@ -2,10 +2,10 @@ const test = require('tape')
 const sinon = require('sinon')
 const { CONFIRM_URL } = require('../../../paths')
 const { handleRequestForPath } = require('./handle-path-request')
-const { states, testUtils } = require('../../state-machine')
+const { states } = require('../../state-machine')
+const { buildSessionForJourney } = require('../../test-utils')
 
 const { IN_PROGRESS, COMPLETED } = states
-const { buildSessionForJourney } = testUtils
 
 const APPLY = 'apply'
