+++ src/web/routes/application/flow-control/state-machine/index.js
@@ -1,11 +1,9 @@
 const { stateMachine } = require('./state-machine')
 const states = require('./states')
 const actions = require('./actions')
-const testUtils = require('./test-utils')
 
 module.exports = {
   states,
   actions,
-  stateMachine,
-  testUtils
+  stateMachine
 }
